TY: Use real std::Result type in try expr and macro
Previously we just compared type name with 'Result'
